requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.26.0
pydantic==2.6.1
python-multipart==0.0.9
Pillow==10.2.0
numpy==1.26.4
chromadb==0.4.24
sentence-transformers==2.4.0
aiofiles==23.2.1
pyyaml==6.0.1
structlog==24.1.0
